% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvWidget.R
\name{igvWidget}
\alias{igvWidget}
\title{igvWidget}
\value{
shiny code which, wnen invoked (almost always by the shinyApp function, returns html

nothing

character string

nothing

a "chrom:start-end" string (or, in rare circumstances, a geneSymbol)

nothing

nothing

nothing

nothing
}
\description{
an R6 class providing clean access to the igvShiny package
}
\details{
R6 Class representing an igv.js genome browser


This class provides the igv.js genome browser in a classic object-oriented-programming R context
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{igvWidget$new()}}
\item \href{#method-ui}{\code{igvWidget$ui()}}
\item \href{#method-server}{\code{igvWidget$server()}}
\item \href{#method-getTrackClickEntity}{\code{igvWidget$getTrackClickEntity()}}
\item \href{#method-setLocus}{\code{igvWidget$setLocus()}}
\item \href{#method-getLocus}{\code{igvWidget$getLocus()}}
\item \href{#method-displayBedTrack}{\code{igvWidget$displayBedTrack()}}
\item \href{#method-displayBedGraphTrack}{\code{igvWidget$displayBedGraphTrack()}}
\item \href{#method-displayVcfTrack}{\code{igvWidget$displayVcfTrack()}}
\item \href{#method-displayGwasTrack}{\code{igvWidget$displayGwasTrack()}}
\item \href{#method-displaySegTrack}{\code{igvWidget$displaySegTrack()}}
\item \href{#method-clone}{\code{igvWidget$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new igvWidget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{igvWidget$new(id, genome, locus, width = "95\%", height = "500px", border)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the html document div id}

\item{\code{genome}}{character string, eg "hg38", "mm10"}

\item{\code{locus}}{character string, a geneSymbol or chrom:start-end}

\item{\code{width}}{integer, in width}

\item{\code{height}}{integer, in pixels}

\item{\code{border}}{character string, for instance "1px solid purple; border-radius: 5px;"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `igvWidget` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui"></a>}}
\if{latex}{\out{\hypertarget{method-ui}{}}}
\subsection{Method \code{ui()}}{
defines the html structure of this widget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{igvWidget$ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-server"></a>}}
\if{latex}{\out{\hypertarget{method-server}{}}}
\subsection{Method \code{server()}}{
defines the event handlers used with the ui
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{igvWidget$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{list, managed by Shiny}

\item{\code{output}}{list, managed by Shiny}

\item{\code{session}}{list (environment?) managed by Shiny}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTrackClickEntity"></a>}}
\if{latex}{\out{\hypertarget{method-getTrackClickEntity}{}}}
\subsection{Method \code{getTrackClickEntity()}}{
igv.js triggers an event when the track is clicked.  we parse that into an
"entity" as best we can:  a gene, a variant, a track block's chromLoc
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{igvWidget$getTrackClickEntity()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setLocus"></a>}}
\if{latex}{\out{\hypertarget{method-setLocus}{}}}
\subsection{Method \code{setLocus()}}{
directs the igv browser to display, and zoom in on, the specified locus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{igvWidget$setLocus(newLocus)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newLocus}}{character string, either a recognized geneSymbol or  "chrom:start-end"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLocus"></a>}}
\if{latex}{\out{\hypertarget{method-getLocus}{}}}
\subsection{Method \code{getLocus()}}{
queries for the currently visible locus in the genome browser
  TODO: need defensive programmer here, different genomes (pshannon, 11 nov 2020)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{igvWidget$getLocus()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-displayBedTrack"></a>}}
\if{latex}{\out{\hypertarget{method-displayBedTrack}{}}}
\subsection{Method \code{displayBedTrack()}}{
displays a BED track in the genome browser
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{igvWidget$displayBedTrack(
  trackName,
  tbl.bed,
  color,
  trackHeight,
  deleteTracksOfSameName = TRUE,
  quiet = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trackName}}{character string, e.g., "ATAC-seq hour 2"}

\item{\code{tbl.bed}}{data.frame with 3 or more columns, typically chrom, start, end, name}

\item{\code{color}}{character string, a standard CSS color name or RGB triple ("#FFAA00") or "random"}

\item{\code{trackHeight}}{integer height in pixels in which to display the track}

\item{\code{deleteTracksOfSameName}}{logical almost always TRUE; FALSE allows multiple tracks with the same name}

\item{\code{quiet}}{logical shows trace of multiple steps along the way}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-displayBedGraphTrack"></a>}}
\if{latex}{\out{\hypertarget{method-displayBedGraphTrack}{}}}
\subsection{Method \code{displayBedGraphTrack()}}{
displays a bedGraph track in the genome browser, uses a numeric scale, score mapped to height
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{igvWidget$displayBedGraphTrack(
  trackName,
  tbl.bg,
  color,
  trackHeight,
  autoscale,
  min = NA_real_,
  max = NA_real_,
  deleteTracksOfSameName = TRUE,
  quiet = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trackName}}{character string, e.g., "ATAC-seq hour 2"}

\item{\code{tbl.bg}}{data.frame with 4 or more columns: chrom, start, end, score}

\item{\code{color}}{character string, a standard CSS color name or RGB triple ("#FFAA00") or "random"}

\item{\code{trackHeight}}{integer height in pixels in which to display the track}

\item{\code{autoscale}}{logical, track score scale calculated from data, or not}

\item{\code{min}}{numeric explicit bound on the track score scale}

\item{\code{max}}{numeric explicit bound on the track score scale}

\item{\code{deleteTracksOfSameName}}{logical almost always TRUE; FALSE allows multiple tracks with the same name}

\item{\code{quiet}}{logical shows trace of multiple steps along the way}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-displayVcfTrack"></a>}}
\if{latex}{\out{\hypertarget{method-displayVcfTrack}{}}}
\subsection{Method \code{displayVcfTrack()}}{
displays a VCF (Variant Call Format) track in the genome browser
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{igvWidget$displayVcfTrack(trackName, vcfData, deleteTracksOfSameName = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trackName}}{character string, e.g., "ATAC-seq hour 2"}

\item{\code{vcfData}}{a largely opaque object read by the VariantAnnotation package from a vcf file}

\item{\code{deleteTracksOfSameName}}{logical almost always TRUE; FALSE allows multiple tracks with the same name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-displayGwasTrack"></a>}}
\if{latex}{\out{\hypertarget{method-displayGwasTrack}{}}}
\subsection{Method \code{displayGwasTrack()}}{
displays a GWAS (genome-wide association study) track in the genome browser
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{igvWidget$displayGwasTrack(trackName, tbl.gwas, deleteTracksOfSameName = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trackName}}{character string, e.g., "ATAC-seq hour 2"}

\item{\code{tbl.gwas}}{data.frame, see details below}

\item{\code{deleteTracksOfSameName}}{logical almost always TRUE; FALSE allows multiple tracks with the same name}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
the GWAS file format has these columns; two example lines follow
              SNPS      cptid CHR_ID  CHR_POS EA NEA  EAF    Beta     SE   P.VALUE Effective_N
 14517  rs79350681  1:5330646      1  5330646  a   g 0.02 -0.7356 0.1875 8.770e-05        2902
 14527 rs187641513  1:5331868      1  5331868  a   g 0.02 -0.7424 0.1889 8.493e-05        2902
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-displaySegTrack"></a>}}
\if{latex}{\out{\hypertarget{method-displaySegTrack}{}}}
\subsection{Method \code{displaySegTrack()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{igvWidget$displaySegTrack(trackName, tbl, deleteTracksOfSameName = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{igvWidget$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
